Create aggregate statistics about info pages.
Gather PP data on all pages with problem reports or search
queries. Aggregate and normalise this data by the number of
unique page views for each page. Push the aggregated data
back to the PP.

This would be run as a regular job. We will then be able to
generate regular reports about the most problematic content,
helping to decide which GOV.UK areas need the most improvement.

There is special handling for smart answers. This is
because GA aggregate page view and search term statistics
across all pages in a smart answer, but Feedex reports
problem reports separately for each page in a smart answer.
It would be good to deal with this when the data is
pushed into the PP.