Test that start and end times are midnight
We want the data to run midnight to midnight, not be dependent on the time the
script is run. This adds tests for this existing behaviour.

This commit also formats the dates correctly before they are set as attributes
on PerformancePlatform, to ensure that they are always used correctly within
the class. This also makes this behaviour easier to test.

If no end_date is passed to InfoStatistics it now defaults to the current time
in UTC; in practice this script will be running on a machine in UTC anyway,
but it's better to be explicit. The expectation that dates will be passed in
UTC is also made explicit in a docstring.