Capture log output from all sources in DEBUG mode
We've got 400 from PP for the POST request, and the client is logging the
response body, but we didn't see it in our logging because we were only using
the logger for the stats package. Using the root logger when the log level is
set to DEBUG will output all logged messages including those from the PP
client, hopefully letting us see why the POST is failing.