Pin all dependencies
requests 2.6.1 errors when trying to parse JSON in the response from PP - this
issue is likely to be the cause:

    https://github.com/kennethreitz/requests/issues/2561

2.6.0 works, so pin to that version. Also pin all other dependencies, for
general safety and consistency.

This way of pinning versions of all dependencies when we have our requirements
split across multiple files is fairly messy and hard to maintain. I've used
pip-compile fairly successfully in the past to manage multiple requirements
files while clearly separating direct dependencies from inferred ones, but it's
still not officially released; it's on a branch of pip-tools and has been for a
long time, although it is being actively worked on at the moment:

    https://github.com/nvie/pip-tools/tree/future

Blog post from 2013 about the plans for pip-tools:

    http://nvie.com/posts/better-package-management/